Move more phase-related functions into phase-manager
