+++ src/data/mystery-encounters/encounters/fun-and-games-encounter.ts
@@ -32,7 +32,7 @@ import { modifierTypes } from "#app/modifier/modifier-type";
 import { Nature } from "#enums/nature";
 import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/game-mode";
 import { isPokemonValidForEncounterOptionSelection } from "#app/data/mystery-encounters/utils/encounter-pokemon-utils";
-import { phaseManager } from "#app/phase-manager";
+import { phaseManager, triggerPokemonFormChange } from "#app/phase-manager";
 
 /** the i18n namespace for the encounter */
 const namespace = "mysteryEncounters/funAndGames";
@@ -400,7 +400,7 @@ function summonPlayerPokemonAnimation(pokemon: PlayerPokemon): Promise<void> {
 
                   pokemon.resetTurnData();
 
-                  globalScene.triggerPokemonFormChange(pokemon, SpeciesFormChangeActiveTrigger, true);
+                  triggerPokemonFormChange(pokemon, SpeciesFormChangeActiveTrigger, true);
                   phaseManager.pushPhase(new PostSummonPhase(pokemon.getBattlerIndex()));
                   resolve();
                 });
