+++ src/phase-manager.ts
@@ -1,7 +1,21 @@
-import type { Phase } from "./phase";
-import { MessagePhase } from "./phases/message-phase";
-import { TurnInitPhase } from "./phases/turn-init-phase";
-import type { Constructor } from "./utils";
+import {
+  type SpeciesFormChangeTrigger,
+  pokemonFormChanges,
+  type SpeciesFormChange,
+  FormChangeItem,
+} from "#app/data/pokemon-forms";
+import type { PokemonAnimType } from "#enums/pokemon-anim-type";
+import { Species } from "#enums/species";
+import type Pokemon from "#app/field/pokemon";
+import { PlayerPokemon } from "#app/field/pokemon";
+import { PokemonFormChangeItemModifier } from "#app/modifier/modifier";
+import type { Phase } from "#app/phase";
+import { FormChangePhase } from "#app/phases/form-change-phase";
+import { MessagePhase } from "#app/phases/message-phase";
+import { PokemonAnimPhase } from "#app/phases/pokemon-anim-phase";
+import { QuietFormChangePhase } from "#app/phases/quiet-form-change-phase";
+import { TurnInitPhase } from "#app/phases/turn-init-phase";
+import type { Constructor } from "#app/utils";
 
 class PhaseManager {
   constructor(
@@ -266,4 +280,70 @@ export function queueMessage(
     //remember that pushPhase adds it to nextCommandPhaseQueue
     phaseManager.pushPhase(phase);
   }
+}
+
+export function triggerPokemonFormChange(
+  pokemon: Pokemon,
+  formChangeTriggerType: Constructor<SpeciesFormChangeTrigger>,
+  delayed: boolean = false,
+  modal: boolean = false,
+): boolean {
+  if (pokemonFormChanges.hasOwnProperty(pokemon.species.speciesId)) {
+    // in case this is NECROZMA, determine which forms this
+    const matchingFormChangeOpts = pokemonFormChanges[pokemon.species.speciesId].filter(
+      (fc) => fc.findTrigger(formChangeTriggerType) && fc.canChange(pokemon),
+    );
+    let matchingFormChange: SpeciesFormChange | null;
+    if (pokemon.species.speciesId === Species.NECROZMA && matchingFormChangeOpts.length > 1) {
+      // Ultra Necrozma is changing its form back, so we need to figure out into which form it devolves.
+      const formChangeItemModifiers = (
+        this.findModifiers(
+          (m) => m instanceof PokemonFormChangeItemModifier && m.pokemonId === pokemon.id,
+        ) as PokemonFormChangeItemModifier[]
+      )
+        .filter((m) => m.active)
+        .map((m) => m.formChangeItem);
+
+      matchingFormChange = formChangeItemModifiers.includes(FormChangeItem.N_LUNARIZER)
+        ? matchingFormChangeOpts[0]
+        : formChangeItemModifiers.includes(FormChangeItem.N_SOLARIZER)
+          ? matchingFormChangeOpts[1]
+          : null;
+    } else {
+      matchingFormChange = matchingFormChangeOpts[0];
+    }
+    if (matchingFormChange) {
+      let phase: Phase;
+      if (pokemon instanceof PlayerPokemon && !matchingFormChange.quiet) {
+        phase = new FormChangePhase(pokemon, matchingFormChange, modal);
+      } else {
+        phase = new QuietFormChangePhase(pokemon, matchingFormChange);
+      }
+      if (pokemon instanceof PlayerPokemon && !matchingFormChange.quiet && modal) {
+        phaseManager.overridePhase(phase);
+      } else if (delayed) {
+        phaseManager.pushPhase(phase);
+      } else {
+        phaseManager.unshiftPhase(phase);
+      }
+      return true;
+    }
+  }
+
+  return false;
+}
+
+export function triggerPokemonBattleAnim(
+  pokemon: Pokemon,
+  battleAnimType: PokemonAnimType,
+  fieldAssets?: Phaser.GameObjects.Sprite[],
+  delayed: boolean = false,
+): boolean {
+  const phase: Phase = new PokemonAnimPhase(battleAnimType, pokemon, fieldAssets);
+  if (delayed) {
+    phaseManager.pushPhase(phase);
+  } else {
+    phaseManager.unshiftPhase(phase);
+  }
+  return true;
 }
